feat: architecture optimization - split graphics, add tests, LRU cache, CI cleanup
