+++ .github/workflows/codecov.yml
@@ -43,6 +43,6 @@ jobs:
       - name: Run tests
         run: pdm run pytest --cov=dynrender_skia --cov-report xml
       - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@v3
+        uses: codecov/codecov-action@v4
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
