+++ dynrender_skia/renderers/__init__.py
@@ -1,21 +1,24 @@
 """Renderer modules."""
 
-from .registry import get_major_renderer, get_additional_renderer
-from .header import BiliHeader, RepostHeader
-from .footer import Footer
-from .text import BiliText
-from .repost import BiliRepost
-
 # Trigger decorator registration for all major and additional types
-from . import major  # noqa: F401
-from . import additional  # noqa: F401
+from . import (
+    additional,  # noqa: F401
+    major,  # noqa: F401
+)
+from .base import safe_run
+from .footer import Footer
+from .header import BiliHeader, RepostHeader
+from .registry import get_additional_renderer, get_major_renderer
+from .repost import BiliRepost
+from .text import BiliText
 
 __all__ = [
     "BiliHeader",
-    "RepostHeader",
-    "Footer",
-    "BiliText",
     "BiliRepost",
-    "get_major_renderer",
+    "BiliText",
+    "Footer",
+    "RepostHeader",
     "get_additional_renderer",
+    "get_major_renderer",
+    "safe_run",
 ]
