+++ tests/conftest.py
@@ -7,8 +7,9 @@ from _pytest.logging import LogCaptureFixture
 from _pytest.tmpdir import TempPathFactory
 from loguru import logger
 
+
 @pytest.fixture(scope="session")
-def shared_cache(tmp_path_factory: TempPathFactory) -> object:
+def shared_cache(tmp_path_factory: TempPathFactory) -> pathlib.Path:
     cache_dir = tmp_path_factory.mktemp("cache", numbered=False)
     logger.info(f"创建共享缓存目录：{cache_dir}")
     yield cache_dir
