+++ tests/test_dynrender_run.py
@@ -1,5 +1,4 @@
 import json
-import os
 
 import aiofiles
 import pytest
@@ -7,9 +6,16 @@ import skia
 from dynamicadaptor.DynamicConversion import formate_message
 
 
-@pytest.mark.skipif(os.getenv("CI") == "true", reason="Skip this test in CI environments.")
+def _has_font(family: str) -> bool:
+    try:
+        return skia.Typeface.MakeFromName(family, skia.FontStyle.Normal()) is not None
+    except Exception:
+        return False
+
+
+@pytest.mark.skipif(not _has_font("Noto Sans SC"), reason="Requires Noto Sans SC font installed")
 @pytest.mark.asyncio
-async def test_dyn_render_run(shared_cache, resource_dir, dynrender_instance):
+async def test_dyn_render_run(resource_dir, dynrender_instance):
     async with aiofiles.open(resource_dir / "message.json", encoding="utf-8") as f:
         resp = await f.read()
